Implement FileSync hardening and modularization updates
