+++ .gitignore
@@ -1,0 +1,1 @@
+filesync/icon.bak.png
