+++ _meta.lua
@@ -1,6 +1,7 @@
 local _ = require("gettext")
 return {
     name = "filesync",
+    version = "1.0.0",
     fullname = _("FileSync - Wireless File Manager"),
     description = _([[Launches a local HTTP web server and displays a QR code. Scan with your phone to manage books and files on your device wirelessly through a clean web interface.]]),
 }
