+++ filesync/fileops_manage.lua
@@ -1,0 +1,329 @@
+return function(FileOps, dependencies)
+    local lfs = dependencies.lfs
+    local logger = dependencies.logger
+
+    --- Create a directory.
+    function FileOps:createDirectory(rel_path, options)
+        local full_path, err = self:_resolvePath(rel_path, options)
+        if not full_path then
+            return false, err
+        end
+
+        local parent = full_path:match("(.+)/[^/]+$")
+        if parent then
+            local parent_attr = lfs.attributes(parent)
+            if not parent_attr or parent_attr.mode ~= "directory" then
+                return false, "Parent directory does not exist"
+            end
+        end
+
+        local attr = lfs.attributes(full_path)
+        if attr then
+            return false, "Path already exists"
+        end
+
+        local dir_name = full_path:match("([^/]+)$")
+        local valid, valid_err = self:_validateFilename(dir_name)
+        if not valid then
+            return false, valid_err
+        end
+
+        local ok, mkdir_err = lfs.mkdir(full_path)
+        if not ok then
+            return false, "Cannot create directory: " .. tostring(mkdir_err)
+        end
+
+        logger.info("FileSync: Created directory", full_path)
+        return true
+    end
+
+    --- Rename a file or directory.
+    function FileOps:rename(old_rel_path, new_rel_path, options)
+        local old_path, err1 = self:_resolvePath(old_rel_path, options)
+        if not old_path then
+            return false, err1
+        end
+
+        local new_path, err2 = self:_resolvePath(new_rel_path, options)
+        if not new_path then
+            return false, err2
+        end
+
+        local attr = lfs.attributes(old_path)
+        if not attr then
+            return false, "Source does not exist"
+        end
+
+        local dest_attr = lfs.attributes(new_path)
+        if dest_attr then
+            return false, "Destination already exists"
+        end
+
+        local new_name = new_path:match("([^/]+)$")
+        local valid, valid_err = self:_validateFilename(new_name)
+        if not valid then
+            return false, valid_err
+        end
+
+        local ok, rename_err = os.rename(old_path, new_path)
+        if not ok then
+            return false, "Cannot rename: " .. tostring(rename_err)
+        end
+
+        logger.info("FileSync: Renamed", old_path, "to", new_path)
+        return true
+    end
+
+    function FileOps:_validateDestinationPath(full_path)
+        local parent = full_path:match("(.+)/[^/]+$")
+        if parent then
+            local parent_attr = lfs.attributes(parent)
+            if not parent_attr or parent_attr.mode ~= "directory" then
+                return false, "Parent directory does not exist"
+            end
+        end
+
+        local name = full_path:match("([^/]+)$")
+        local valid, valid_err = self:_validateFilename(name)
+        if not valid then
+            return false, valid_err
+        end
+
+        return true
+    end
+
+    function FileOps:_isPathInside(parent_path, child_path)
+        return child_path:sub(1, #parent_path + 1) == parent_path .. "/"
+    end
+
+    --- Move a file or directory.
+    function FileOps:move(old_rel_path, new_rel_path, options)
+        options = options or {}
+        local old_path, err1, old_scope = self:_resolvePath(old_rel_path, {
+            scope = options.old_scope or options.scope,
+            allow_root_scopes = options.allow_root_scopes,
+        })
+        if not old_path then
+            return false, err1
+        end
+
+        local new_path, err2, new_scope = self:_resolvePath(new_rel_path, {
+            scope = options.new_scope or options.scope,
+            allow_root_scopes = options.allow_root_scopes,
+        })
+        if not new_path then
+            return false, err2
+        end
+
+        if old_path == old_scope.root_path then
+            return false, "Cannot move root directory"
+        end
+
+        local src_attr = lfs.attributes(old_path)
+        if not src_attr then
+            return false, "Source does not exist"
+        end
+
+        local dest_attr = lfs.attributes(new_path)
+        if dest_attr then
+            return false, "Destination already exists"
+        end
+
+        local valid, valid_err = self:_validateDestinationPath(new_path)
+        if not valid then
+            return false, valid_err
+        end
+
+        if src_attr.mode == "directory" and old_scope.id == new_scope.id and self:_isPathInside(old_path, new_path) then
+            return false, "Cannot move directory inside itself"
+        end
+
+        local ok, move_err = os.rename(old_path, new_path)
+        if not ok then
+            return false, "Cannot move: " .. tostring(move_err)
+        end
+
+        logger.info("FileSync: Moved", old_path, "to", new_path)
+        return true
+    end
+
+    --- Copy a file in chunks without loading it entirely into memory.
+    function FileOps:copyFile(src_rel_path, dst_rel_path, options)
+        options = options or {}
+        local src_path, err1 = self:_resolvePath(src_rel_path, {
+            scope = options.old_scope or options.scope,
+            allow_root_scopes = options.allow_root_scopes,
+        })
+        if not src_path then
+            return false, err1
+        end
+
+        local dst_path, err2 = self:_resolvePath(dst_rel_path, {
+            scope = options.new_scope or options.scope,
+            allow_root_scopes = options.allow_root_scopes,
+        })
+        if not dst_path then
+            return false, err2
+        end
+
+        local src_attr = lfs.attributes(src_path)
+        if not src_attr then
+            return false, "Source does not exist"
+        end
+        if src_attr.mode ~= "file" then
+            return false, "Copy only supports files"
+        end
+
+        local dest_attr = lfs.attributes(dst_path)
+        if dest_attr then
+            return false, "Destination already exists"
+        end
+
+        local valid, valid_err = self:_validateDestinationPath(dst_path)
+        if not valid then
+            return false, valid_err
+        end
+
+        local src_file, src_err = io.open(src_path, "rb")
+        if not src_file then
+            return false, "Cannot open source file: " .. tostring(src_err)
+        end
+
+        local dst_file, dst_err = io.open(dst_path, "wb")
+        if not dst_file then
+            src_file:close()
+            return false, "Cannot create destination file: " .. tostring(dst_err)
+        end
+
+        local chunk_size = 65536
+        while true do
+            local chunk, read_err = src_file:read(chunk_size)
+            if read_err then
+                src_file:close()
+                dst_file:close()
+                os.remove(dst_path)
+                return false, "Cannot read source file: " .. tostring(read_err)
+            end
+            if not chunk then break end
+            local write_ok, write_err = dst_file:write(chunk)
+            if not write_ok then
+                src_file:close()
+                dst_file:close()
+                os.remove(dst_path)
+                return false, "Cannot write destination file: " .. tostring(write_err)
+            end
+        end
+
+        local close_ok, close_err = dst_file:close()
+        src_file:close()
+        if close_ok == false then
+            os.remove(dst_path)
+            return false, "Cannot finalize destination file: " .. tostring(close_err)
+        end
+
+        logger.info("FileSync: Copied", src_path, "to", dst_path)
+        return true
+    end
+
+    --- Delete a file or directory (directory must be empty unless recursive=true).
+    function FileOps:delete(rel_path, options)
+        options = options or {}
+        local full_path, err, scope = self:_resolvePath(rel_path, options)
+        if not full_path then
+            return false, err
+        end
+
+        if full_path == scope.root_path then
+            return false, "Cannot delete root directory"
+        end
+
+        local attr = lfs.attributes(full_path)
+        if not attr then
+            return false, "Path does not exist"
+        end
+
+        local is_file = attr.mode ~= "directory"
+
+        if attr.mode == "directory" then
+            local child_count = 0
+            for child_name in lfs.dir(full_path) do
+                if child_name ~= "." and child_name ~= ".." then
+                    child_count = child_count + 1
+                    break
+                end
+            end
+
+            if child_count == 0 then
+                local ok, del_err = lfs.rmdir(full_path)
+                if not ok then
+                    return false, "Cannot delete directory: " .. tostring(del_err)
+                end
+            elseif options.recursive == true then
+                local ok, del_err = self:_deleteRecursive(full_path)
+                if not ok then
+                    return false, "Cannot delete directory: " .. tostring(del_err)
+                end
+            else
+                return false, "Cannot delete non-empty directory without recursive flag"
+            end
+        else
+            local ok, del_err = os.remove(full_path)
+            if not ok then
+                return false, "Cannot delete file: " .. tostring(del_err)
+            end
+        end
+
+        logger.info("FileSync: Deleted", full_path)
+
+        if is_file and options then
+            local should_delete_sdr = false
+            if options.safe_mode then
+                should_delete_sdr = true
+            elseif options.delete_sdr then
+                should_delete_sdr = true
+            end
+
+            if should_delete_sdr then
+                local sdr_path = full_path .. ".sdr"
+                local sdr_attr = lfs.attributes(sdr_path)
+                if sdr_attr and sdr_attr.mode == "directory" then
+                    local sdr_ok, sdr_err = self:_deleteRecursive(sdr_path)
+                    if sdr_ok then
+                        logger.info("FileSync: Deleted .sdr metadata directory", sdr_path)
+                    else
+                        logger.warn("FileSync: Failed to delete .sdr directory", sdr_path, sdr_err)
+                    end
+                end
+            end
+        end
+
+        return true
+    end
+
+    --- Recursively delete a directory and its contents.
+    function FileOps:_deleteRecursive(path)
+        for name in lfs.dir(path) do
+            if name ~= "." and name ~= ".." then
+                local entry_path = path .. "/" .. name
+                local entry_attr = lfs.attributes(entry_path)
+                if entry_attr then
+                    if entry_attr.mode == "directory" then
+                        local ok, err = self:_deleteRecursive(entry_path)
+                        if not ok then return false, err end
+                    else
+                        local ok, err = os.remove(entry_path)
+                        if not ok then
+                            return false, "Cannot delete: " .. tostring(err)
+                        end
+                    end
+                end
+            end
+        end
+
+        local ok, err = lfs.rmdir(path)
+        if not ok then
+            return false, "Cannot remove directory: " .. tostring(err)
+        end
+        return true
+    end
+end
